+++ R/optimACDC.R
@@ -147,6 +147,18 @@ optimACDC <-
                              n.pts = n_pts, n.cov = n_cov, utopia = utopia,
                              covars.type = covars.type)
       
+      # Avoid the following error:
+      # Error in if (new_energy[1] <= old_energy[1]) { : 
+      #   missing value where TRUE/FALSE needed
+      # Source: http://stackoverflow.com/a/7355280/3365410
+      # ASR: The reason for the error is unknown to me.
+      if (is.na(new_energy)) {
+        new_energy <- old_energy
+        new_conf <- old_conf
+        new_sm <- old_sm
+        new_scm <- old_scm
+      }
+      
       # Evaluate the new system configuration
       if (greedy) {
         random_prob <- 1