+++ R/optimCORR.R
@@ -101,6 +101,18 @@ optimCORR <-
       new_scm <- .corCORR(obj = new_sm, covars.type = covars.type)
       new_energy <- .objCORR(scm = new_scm, pcm = pcm)
       
+      # Avoid the following error:
+      # Error in if (new_energy <= old_energy) { : 
+      #   missing value where TRUE/FALSE needed
+      # Source: http://stackoverflow.com/a/7355280/3365410
+      # ASR: The reason for the error is unknown to me.
+      if (is.na(new_energy)) {
+        new_energy <- old_energy
+        new_conf <- old_conf
+        new_sm <- old_sm
+        new_scm <- old_scm
+      }
+      
       # Evaluate the new system configuration
       if (greedy) {
         random_prob <- 1