+++ R/spsann.R
@@ -30,8 +30,8 @@
 #' \tabular{ll}{
 #' Package: \tab spsann    \cr
 #' Type:    \tab Package   \cr
-#' Version: \tab 1.0.1     \cr
-#' Date:    \tab 2015-07-14\cr
+#' Version: \tab 1.0-2     \cr
+#' Date:    \tab 2015-08-19\cr
 #' License: \tab GPL (>= 2)\cr
 #' }
 #'